Implementar configurações de usuário, planos e FAQ com melhorias na experiência
- Adicionar view de configuração do usuário com upload de avatar e gerenciamento de dados pessoais
- Criar view de planos com opções de assinatura e FAQs relacionados
- Criar view de FAQ do site para dúvidas gerais
- Melhorar view de perfil, corrigindo pequenos bugs e aprimorando a experiência do usuário
- Reestruturar template base e adicionar notificações tipo toast
- Aprimorar estilos CSS e animações para toasts
- Remover imagens de notificação antigas e adicionar novas para melhor engajamento
- Refatorar index.php para melhorar roteamento e autenticação

+++ lima/Projetos/TechFit/app/models/Aviso.php
@@ -34,6 +34,9 @@ class Aviso
 
         $avisosPath = $stmt->fetchAll(PDO::FETCH_ASSOC);
         foreach ($avisosPath as $avisos) {
+            if(empty($avisos['anexo_path'])) {
+                continue;
+            }
             $filename = __DIR__ . "/../../public/" . $avisos['anexo_path'];
             if (file_exists($filename)) {
                 if (unlink($filename)) {
